+++ src/constant.ts
@@ -31,3 +31,171 @@ export const SKLAND_SM_CONFIG = {
   apiHost: 'fp-it.portal101.cn',
   apiPath: '/deviceprofile/v4',
 }
+
+// DES_RULE 的类型定义
+export interface DESRule {
+  cipher?: string
+  is_encrypt: number
+  key?: string
+  obfuscated_name: string
+}
+
+export const DES_RULE: Record<string, DESRule> = {
+  appId: {
+    cipher: 'DES',
+    is_encrypt: 1,
+    key: 'uy7mzc4h',
+    obfuscated_name: 'xx',
+  },
+  box: {
+    is_encrypt: 0,
+    obfuscated_name: 'jf',
+  },
+  canvas: {
+    cipher: 'DES',
+    is_encrypt: 1,
+    key: 'snrn887t',
+    obfuscated_name: 'yk',
+  },
+  clientSize: {
+    cipher: 'DES',
+    is_encrypt: 1,
+    key: 'cpmjjgsu',
+    obfuscated_name: 'zx',
+  },
+  organization: {
+    cipher: 'DES',
+    is_encrypt: 1,
+    key: '78moqjfc',
+    obfuscated_name: 'dp',
+  },
+  os: {
+    cipher: 'DES',
+    is_encrypt: 1,
+    key: 'je6vk6t4',
+    obfuscated_name: 'pj',
+  },
+  platform: {
+    cipher: 'DES',
+    is_encrypt: 1,
+    key: 'pakxhcd2',
+    obfuscated_name: 'gm',
+  },
+  plugins: {
+    cipher: 'DES',
+    is_encrypt: 1,
+    key: 'v51m3pzl',
+    obfuscated_name: 'kq',
+  },
+  pmf: {
+    cipher: 'DES',
+    is_encrypt: 1,
+    key: '2mdeslu3',
+    obfuscated_name: 'vw',
+  },
+  protocol: {
+    is_encrypt: 0,
+    obfuscated_name: 'protocol',
+  },
+  referer: {
+    cipher: 'DES',
+    is_encrypt: 1,
+    key: 'y7bmrjlc',
+    obfuscated_name: 'ab',
+  },
+  res: {
+    cipher: 'DES',
+    is_encrypt: 1,
+    key: 'whxqm2a7',
+    obfuscated_name: 'hf',
+  },
+  rtype: {
+    cipher: 'DES',
+    is_encrypt: 1,
+    key: 'x8o2h2bl',
+    obfuscated_name: 'lo',
+  },
+  sdkver: {
+    cipher: 'DES',
+    is_encrypt: 1,
+    key: '9q3dcxp2',
+    obfuscated_name: 'sc',
+  },
+  status: {
+    cipher: 'DES',
+    is_encrypt: 1,
+    key: '2jbrxxw4',
+    obfuscated_name: 'an',
+  },
+  subVersion: {
+    cipher: 'DES',
+    is_encrypt: 1,
+    key: 'eo3i2puh',
+    obfuscated_name: 'ns',
+  },
+  svm: {
+    cipher: 'DES',
+    is_encrypt: 1,
+    key: 'fzj3kaeh',
+    obfuscated_name: 'qr',
+  },
+  time: {
+    cipher: 'DES',
+    is_encrypt: 1,
+    key: 'q2t3odsk',
+    obfuscated_name: 'nb',
+  },
+  timezone: {
+    cipher: 'DES',
+    is_encrypt: 1,
+    key: '1uv05lj5',
+    obfuscated_name: 'as',
+  },
+  tn: {
+    cipher: 'DES',
+    is_encrypt: 1,
+    key: 'x9nzj1bp',
+    obfuscated_name: 'py',
+  },
+  trees: {
+    cipher: 'DES',
+    is_encrypt: 1,
+    key: 'acfs0xo4',
+    obfuscated_name: 'pi',
+  },
+  ua: {
+    cipher: 'DES',
+    is_encrypt: 1,
+    key: 'k92crp1t',
+    obfuscated_name: 'bj',
+  },
+  url: {
+    cipher: 'DES',
+    is_encrypt: 1,
+    key: 'y95hjkoo',
+    obfuscated_name: 'cf',
+  },
+  version: {
+    is_encrypt: 0,
+    obfuscated_name: 'version',
+  },
+  vpw: {
+    cipher: 'DES',
+    is_encrypt: 1,
+    key: 'r9924ab5',
+    obfuscated_name: 'ca',
+  },
+}
+
+export const BROWSER_ENV = {
+  plugins: 'MicrosoftEdgePDFPluginPortableDocumentFormatinternal-pdf-viewer1,MicrosoftEdgePDFViewermhjfbmdgcfjbbpaeojofohoefgiehjai1',
+  ua: 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/129.0.0.0 Safari/537.36 Edg/129.0.0.0',
+  canvas: '259ffe69', // 基于浏览器的canvas获得的值，不知道复用行不行
+  timezone: -480, // 时区，应该是固定值吧
+  platform: 'Win32',
+  url: 'https://www.skland.com/', // 固定值
+  referer: '',
+  res: '1920_1080_24_1.25', // 屏幕宽度_高度_色深_window.devicePixelRatio
+  clientSize: '0_0_1080_1920_1920_1080_1920_1080',
+  status: '0011', // 不知道在干啥
+}
