+++ src/utils.ts
@@ -1,7 +1,9 @@
-import { createHash, createHmac } from 'node:crypto'
 import type { FetchContext } from 'ofetch'
 import { stringifyQuery } from 'ufo'
-import { get_d_id } from './did'
+import pako from 'pako'
+import { format } from 'date-fns'
+import { BROWSER_ENV, DES_RULE, SKLAND_SM_CONFIG } from './constant'
+import { encryptAES, encryptObjectByDESRules, encryptRSA, hmacSha256, md5 } from './crypto'
 
 export const command_header = {
   'User-Agent': 'Skland/1.21.0 (com.hypergryph.skland; build:102100065; iOS 17.6.0; ) Alamofire/5.7.1',
@@ -34,7 +36,7 @@ export function getRequestURL(request: RequestInfo, baseURL?: string) {
 
 const WHITE_LIST = ['/web/v1/user/auth/generate_cred_by_code']
 
-export function onSignatureRequest(ctx: FetchContext) {
+export async function onSignatureRequest(ctx: FetchContext) {
   const { pathname } = getRequestURL(ctx.request, ctx.options.baseURL)
 
   if (WHITE_LIST.includes(pathname))
@@ -55,13 +57,9 @@ export function onSignatureRequest(ctx: FetchContext) {
   }
   const str = `${pathname}${query}${ctx.options.body ? JSON.stringify(ctx.options.body) : ''}${timestamp}${JSON.stringify(signatureHeaders)}`
 
-  const hmacSha256ed = createHmac('sha256', token)
-    .update(str, 'utf-8')
-    .digest('hex')
+  const hmacSha256ed = await hmacSha256(token, str)
 
-  const sign = createHash('md5')
-    .update(hmacSha256ed)
-    .digest('hex')
+  const sign = await md5(hmacSha256ed)
 
   Object.entries(signatureHeaders).forEach(([key, value]) => {
     headers.append(key, value)
@@ -72,6 +70,136 @@ export function onSignatureRequest(ctx: FetchContext) {
   ctx.options.headers = headers
 }
 
-export function createDeviceId() {
-  return get_d_id()
+const stringify = (obj: any) => JSON.stringify(obj).replace(/":"/g, '": "').replace(/","/g, '", "')
+
+export function gzipObject(o: object) {
+  const jsonStr = stringify(o)
+  const encoded = new TextEncoder().encode(jsonStr)
+  const compressed = pako.gzip(encoded, {
+    level: 2,
+  })
+
+  // Python gzip OS FLG = Unknown
+  compressed.set([19], 9)
+
+  return btoa(String.fromCharCode(...compressed))
+}
+
+export async function getSmId() {
+  const now = new Date()
+  const _time = format(now, 'yyyyMMddHHmmss')
+
+  // 生成UUID v4
+  const uid = crypto.randomUUID()
+
+  // MD5加密uid
+  const uidMd5 = md5(uid)
+
+  const v = `${_time + uidMd5}00`
+
+  // 计算smsk_web
+  const smsk_web = (await md5(`smsk_web_${v}`)).substring(0, 14)
+
+  return `${v + smsk_web}0`
+}
+
+export function getTn(o: Record<string, any>) {
+  // 获取并排序对象的所有键
+  const sortedKeys: string[] = Object.keys(o).sort()
+
+  // 用于存储处理后的值
+  const resultList: string[] = []
+
+  // 遍历排序后的键
+  for (const key of sortedKeys) {
+    let v: any = o[key]
+
+    // 处理数字类型
+    if (typeof v === 'number')
+      v = String(v * 10000)
+
+    // 处理对象类型（递归）
+    else if (typeof v === 'object' && v !== null)
+      v = getTn(v)
+
+    resultList.push(v)
+  }
+
+  // 将所有结果连接成字符串
+  return resultList.join('')
+}
+
+const SM_CONFIG = SKLAND_SM_CONFIG
+const devices_info_url = `${SKLAND_SM_CONFIG.protocol}://${SKLAND_SM_CONFIG.apiHost}${SKLAND_SM_CONFIG.apiPath}`
+
+export async function getDid() {
+  // 生成 UUID 并计算 priId
+  const uid = crypto.randomUUID()
+  const priId = (await md5(uid)).substring(0, 16)
+
+  const ep = await encryptRSA(uid, SM_CONFIG.publicKey)
+
+  // 准备浏览器环境数据
+  const browser = {
+    ...BROWSER_ENV,
+    vpw: crypto.randomUUID(),
+    svm: Date.now(),
+    trees: crypto.randomUUID(),
+    pmf: Date.now(),
+  }
+
+  // 准备加密目标数据
+  const desTarget: Record<string, string | number> = {
+    ...browser,
+    protocol: 102,
+    organization: SM_CONFIG.organization,
+    appId: SM_CONFIG.appId,
+    os: 'web',
+    version: '3.0.0',
+    sdkver: '3.0.0',
+    box: '', // 首次请求为空
+    rtype: 'all',
+    smid: await getSmId(),
+    subVersion: '1.0.0',
+    time: 0,
+  }
+
+  // 计算并添加 tn
+  desTarget.tn = await md5(getTn(desTarget))
+
+  // DES 加密
+  const desResult = await encryptObjectByDESRules(desTarget, DES_RULE)
+
+  // GZIP 压缩
+  const gzipResult = gzipObject(desResult)
+
+  // AES 加密
+  const aesResult = await encryptAES(gzipResult, priId)
+
+  const body = {
+    appId: 'default',
+    compress: 2,
+    data: aesResult,
+    encode: 5,
+    ep,
+    organization: SM_CONFIG.organization,
+    os: 'web',
+  }
+
+  // 发送请求
+  const response = await fetch(devices_info_url, {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json',
+    },
+    body: JSON.stringify(body),
+  })
+
+  const resp = await response.json()
+  if (resp.code !== 1100) {
+    console.log(resp)
+    throw new Error('did计算失败，请联系作者')
+  }
+
+  return `B${resp.detail.deviceId}`
 }
